requirements.txt
numpy
scipy
pyyaml
tqdm
matplotlib
opencv-python
ultralytics
torch>=2.0.0
torchvision>=0.15.0
hydra-core
iopath
decord
einops
timm
